### 2025-06-27 #714 @simpar1471
openFDA: 'openFDA' API
Submitting Author Name: Simon Parker
Submitting Author Github Handle: @simpar1471
Repository:  https://www.github.com/simpar1471/openFDA
Version submitted: 0.1.0.9000
Submission type: Standard

Version accepted: TBD

---

```
Package: openFDA
Title: 'openFDA' API
Version: 0.1.0.9000
    person("Simon", "Parker", , "simon.parker.24@ucl.ac.uk",
Description: The 'openFDA' API facilitates access to U.S. Food and Drug
    Administration (FDA) data on drugs, devices, foodstuffs, tobacco, and more.
    This package makes the API easily accessible with 'httr2', returning
    objects which the user can convert to JSON data and parse. Kass-Hout TA, Xu
    Z, Mohebbi M et al. (2016) <doi:10.1093/jamia/ocv153>.
    https://simpar1471.github.io/openFDA/
    cli,
    httr2,
    checkmate,
    vctrs
Config/Needs/website: rmarkdown
    stringr
```

`openFDA` retrieves data from the elastic-search based openFDA API, using `httr2`. The package can be used to query the API on different fields, and successful queries return `httr2` response objects with JSON data that R users can manipulate further.

The package is intended for anyone wanting to use the openFDA API through R. These include researchers focused on pharmacovigilance (e.g. with the device adverse event endpoints) or veterinary researchers (e.g. with the animal product adverse event endpoints) - really, any user interested in accessing FDA data.

There are other packages. These include the [rOpenHealth/openfda](https://github.com/rOpenHealth/openfda) package, in which queries are built up in an `httr2`-like manner. It automatically converts the response data to data frames, but this package returns the underlying `httr2` response object, which users can then choose how to parse themselves. Their package does not have a CI suite, nor a documentation website. There is also [ck2136/FDAopenR](https://github.com/ck2136/FDAopenR/), which has a test suite with 100% coverage and a now-defunct link to Travis CI. It takes an object-oriented approach which I haven't quite wrapped my head around from looking at the README.

Personally, I prefer 'one function, more parameters' approach in my package - but to each their own. I also believe that this package has a wider and more useful array of error handlers, which point the user to commonly encountered problems with the openFDA API itself. This package is also cleared by the relatively more stringent `pkgcheck()` checks.

This package is also already on CRAN. I know the guide for authors specifies you'd rather not have it this way round, but I didn't intend to submit the package for review when I first made it.

Everything passes on CI. My functions `get_api_key()` and `set_api_key()` are not unique to this package, but I think the names are appropriate to the actions these functions perform.
