### 2025-06-27 #714 @simpar1471
Re some findings from the automated checks:

- For `pkgcheck` note on unused imports: The function `vctrs::list_drop_empty()` is used in `openFDA()`, so needs to be imported.
- The `lintr` issue with using `library()` or `require()` I think is coming from the vignettes; there are no `library()`/`require()` calls in the package code itself.
