### 2025-07-02 #714 @mpadge
Thanks for your submission @simpar1471. And thanks for explaining those automated checks, but you don't need to worry. The most important checks are the ✔ ones, which in your case indicate that all is good. Thank you also for the links to equivalent packages, from which it's immediately clear that https://github.com/rOpenHealth/openfda should potentially be archived, which we'll now definitely follow up on. But that's also very helpful to indicate the your package is clearly in scope here, and will likely serve as a great replacement and update for the 'openfda' package. I'll ask for further thoughts from the editorial team, and return either with any questions we may have, or by simply assigning an editor to handle the submission.

Oh, and it's also okay to already have the package on CRAN. The review-then-CRAN sequence is only recommended, as packages often undergo extensive changes during review.
