### 2025-07-03 #714 @mpadge
(And apologies there, I mis-read https://github.com/rOpenHealth/openfda as `ropensci/openfda`, but it's actually nothing to do with us, so I guess we can just kinda ignore that and proceed here?)
