### 2025-07-03 2ff13c4
feat: Added `paging` variable as a package-level option and parameter in `openFDA()`, along with relevant unit tests
