+++ tests/testthat/test-openFDA.R
@@ -1,5 +1,6 @@
 encrypted_api_key <-
   "TEaDtqdFMq9_Montij5p9IY6T57IyqkbF8IYFVOpk-ttxotFUNdJSxgccAnkq4nQhplaf-r3deQ"
+rlang::push_options(openFDA.paging = "no-quiet")
 
 test_that("openFDA can call its API", {
   set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
@@ -7,7 +8,7 @@ test_that("openFDA can call its API", {
   # Simple query
   resp <- openFDA(
     search = c("openfda.brand_name" = "ozempic"),
-    limit = 1
+    limit = 5
   )
   expect_s3_class(resp, "httr2_response")
 })
@@ -25,14 +26,38 @@ test_that("All openFDA endpoints are valid", {
                  "other-historicaldocument", "other-nsde",
                  "other-substance", "other-unii", "tobacco-problem")
   purrr::map(endpoints, function(endpoint) {
-    resp <- expect_no_warning(openFDA(search = "",
-                                      endpoint = endpoint,
-                                      limit = 1,
-                                      paging = "no-quiet"))
+    resp <- expect_no_warning(suppressMessages(
+      openFDA(search = "", endpoint = endpoint, limit = 1, paging = "no-quiet")
+    ))
     expect_s3_class(resp, class = "httr2_response")
   })
 })
 
+test_that("openFDA paging is possible", {
+  set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
+
+  # Query which requires paging
+  resps <- openFDA(
+    search = "openfda.brand_name:o*",
+    limit = 300,
+    paging = "yes-quiet"
+  )
+
+  purrr::map(resps, \(resp) expect_s3_class(resp, "httr2_response"))
+})
+
+test_that(
+  desc = "openFDA paging fails if `paging == \"ask\" but not interactive", {
+  skip_if_not(!interactive())
+  set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
+
+    # Query which requires paging
+    expect_error(
+      openFDA(search = "openfda.brand_name:o*", limit = 300, paging = "ask"),
+      class = "openfda_paging_is_ask_when_uninteractive"
+    )
+  }
+)
 
 test_that("openFDA throws formatted HTTP errors", {
   set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
@@ -88,6 +113,20 @@ test_that("openFDA throws formatted HTTP errors", {
     class ="openFDA_http_error_500"
   )
   expect_s3_class(resp, "httr2_response")
+
+  # Mock up a response with a non-customised error messaage
+  temporary_mock <- function(req) {
+    httr2::response_json(status_code = 503)
+  }
+  expect_warning(
+    httr2::with_mocked_responses(temporary_mock, {
+      openFDA(
+        search = "openfda.brand_name:\"verapamil\"",
+        limit = 1
+      )
+    }),
+    class ="openFDA_http_error_503"
+  )
 })
 
 test_that("openFDA errors on certain bad inputs", {
@@ -166,9 +205,30 @@ test_that("openFDA errors on certain bad inputs", {
     class = "openFDA_skip_too_large"
   )
 
+  ## paging
+  expect_error(
+    openFDA(search = c("openfda.brand_name" = "ozempic"),
+            paging = 1111111),
+    class = "openFDA_invalid_string_param_class"
+  )
+  expect_error(
+    openFDA(search = c("openfda.brand_name" = "ozempic"),
+            paging = "bad-value"),
[…]
