### 2025-07-03 1cf671c
chore: add paging option to _pkgdown.yml
