### 2025-07-03 e505a98
chore: update .Rbuildignore and DESCRIPTION based on R CMD CHECK issues
