### 2025-07-03 8c58f22
docs: update package articles/vignettes
