+++ vignettes/openFDA.Rmd
@@ -32,14 +32,15 @@ knitr::opts_chunk$set(
 )
 ```
 
-Using `{openFDA}` can be a breeze, if you know how to construct good queries.
+Using openFDA can be a breeze, if you know how to construct good queries.
 This short guide will get you started with `{openFDA}`, and show you how to put together more complicated queries.
+We start by loading openFDA.
 
 ```{r load_openFDA}
 library(openFDA)
 ```
 
-```{r set_key}
+```{r set_key, echo = FALSE}
 set_api_key("tvzdUXoC3Yi21hhxtKKochJlQgm6y1Lr7uhmjcpT")
 ```
 
@@ -47,14 +48,36 @@ set_api_key("tvzdUXoC3Yi21hhxtKKochJlQgm6y1Lr7uhmjcpT")
 The openFDA API makes public FDA data available from a simple, public API.
 Users of this API have access to FDA data on food, human and veterinary drugs, devices, and more.
 You can read all about it at their [website](https://open.fda.gov/).
+The API limits results sets to 1000, and pages through results if a user submits a query with >1000 results.
+By default, `openFDA()` will ask you if you want to perform paging, but this tends to break vignettes.
+Therefore, we set the option `openFDA.paging` to `"yes"` in order to automatically perform paging when needed.
+
+```{r set-paging, echo = FALSE}
+options("openFDA.paging" = "no-quiet")
+```
 
 # A simple openFDA query
 The simplest way to query the openFDA API is to identify the **endpoint** you want to use and provide other search terms.
-For example, this snippet retrieves 1 record about **adverse events** in the **drugs** endpoint.
+For example, this snippet retrieves 2 records about **adverse events** in the **drugs** endpoint.
 The empty search string (`""`) means the results will be non-specific.
 
 ```{r example1}
-search <- openFDA(search = "", endpoint = "drug-event", limit = 1)
+search <- openFDA(search = "",
+                  endpoint = "drug-event",
+                  limit = 2,
+                  paging = "no")
+search
+```
+
+The warning given about pagination is because we searched on the whole dataset, but used `limit` to only pull out the first two results.
+We can suppress this warning by setting `paging` to `"no-quiet"`.
+This ensures you don't page the whole dataset (which would take eleven hours) and suppresses the warning:
+
+```{r example2}
+search <- openFDA(search = "",
+                  endpoint = "drug-event",
+                  limit = 2,
+                  paging = "no-quiet")
 search
 ```
 
@@ -62,7 +85,7 @@ search
 The function returns an `{httr2}` response object, with attached JSON data.
 We use `httr2::resp_body_json()` to extract the underlying data.
 
-```{r example2_json}
+```{r example_3a_json}
 json <- httr2::resp_body_json(search)
 ```
 
@@ -73,12 +96,12 @@ The `meta` section has important metadata on your results, which includes:
 
 * `disclaimer` - An important disclaimer regarding the data provided by openFDA.
 * `license` - A webpage with license terms that govern the openFDA API.
-* `last_updated` - The last date when this openFDA endpoint was updated.
+* `last_updated` - The last date at which this openFDA endpoint was updated.
 * `results.skip` - How many results were skipped? Set by the `skip` parameter in `openFDA()`.
 * `results.limit` - How many results were retrieved? Set by the `limit` parameter in `openFDA()`.
 * `results.total` - How many results were there in total matching your `search` criteria?
 
-```{r example_2a_json_meta}
+```{r example_3b_json_meta}
 json$meta
 ```
 
@@ -86,7 +109,7 @@ json$meta
 
 For non-`count` queries, this will be a set of records which were found in the endpoint and match your `search` term.
 
-```{r example_2b_json_results, out.lines = 10}
[…]
