### 2025-07-03 4c9fedd
docs: update examples for R CMD CHECK
