+++ R/openFDA.R
@@ -111,7 +111,8 @@
 #'
 #'   resp <- openFDA(search = "openfda.manufacturer_name:gilead*",
 #'                   limit = 2,
-#'                   skip = 10)
+#'                   skip = 10,
+#'                   paging = "no-quiet")
 #'
 #'   # The function returns an `httr2` object
 #'   print(resp)
