### 2025-07-03 4fbb323
docs: update NEWS.md
