### 2025-07-03 1e1b9cf
chore: bump version to reflect dev changes and update codemeta.json to reflect current repo state
