### 2025-07-03 3304254
fix: fix CI R CMD CHECK issues with downlit syntax in roxygen
