### 2025-07-03 e9ad555
fix: Move test for bad `paging` variable to its own test_that() block, as it's failing on CI. Maybe a new R version issue?
