### 2025-07-03 3353ff7
fix: Fix `paging` variable test. The solution was to use the actualy function argument for the regular expression, `regexp` - not `pattern`.
