+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   openFDA: openFDA.html
   articles/using_openFDA: using_openFDA.html
-last_built: 2025-07-03T22:49Z
+last_built: 2025-07-03T23:17Z
 urls:
   reference: https://simpar1471.github.io/openFDA/reference
   article: https://simpar1471.github.io/openFDA/articles
