### 2025-07-03 3a1921e
fix: add OPENFDA_KEY env var to pkgcheck CI workflow
