### 2025-07-03 e2f59d7
fix: change @name to @rdname - might prevent pkgcheck issue
