### 2025-07-04 #714 @maelle
Thanks a lot for your submission @simpar1471! I have some requests/comments that we can discuss before I look for reviewers. Some of them, especially the ones about the interface, could be left for later once you've heard from reviewers as they might have different opinions and ideas.

---

- In the DESCRIPTION file there's the DOI to a publication about the API. Could the text make clearer what the DOI is?
- The most recent pkgcheck CI runs fail: https://github.com/simpar1471/openFDA/actions?query=workflow%3Apkgcheck

> These functions do not have examples: [openFDA_options].
> The following function has no documented return value: [openFDA_options]

- The docs state "To permanently set the API key for a given project, set `OPENFDA_TOKEN` in `.Renviron`." I would actually recommend mentioning the use of the [keyring package](https://blog.r-hub.io/2024/02/28/key-advantages-of-using-keyring/). Either having it in the docs, or even using it under the hood.

- The `set_api_key()` could instead take no argument, or support not supplying an argument, and then read the key interactively, using `readline()` like in [gitcreds](https://github.com/r-lib/gitcreds/blob/d56bea0dd1f29f228951cf1824d9e66a9457aad6/R/git-auth.R#L111). Alternatively, if you rely on keyring there'll be a prompt. The advantage of having users enter the key alternatively is that it won't be in their R history or scripts.

- This line [if (api_key == "") throw_api_key_error()](https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/api_key.R#L35-L36) could be

```r
if (!nzchar(api_key)) throw_api_key_error()
```

see the docs of the corresponding lintr: <https://lintr.r-lib.org/reference/nzchar_linter.html>

- Very minor: This [line](https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/README.Rmd#L29-L30) should use `--` instead of `-`.

- In the README I see "With an API key you can make 240 requests per minute, and are limited to 120,000 requests per day.". I think it'd be worth mentioning that your package automatically handles the request rate limiting / throttling.

- In https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/api_key.R#L45-L52

```r
  cli::cli_abort(
    c("To use {.pkg openFDA}, you must set an openFDA API key.",
      "i" = paste0("Go to <https://open.fda.gov/apis/authentication/> ",
                   "to get an openFDA API key, then supply it to ",
                   "{.fun set_api_key} to cache it for use in this session.")),
    class = "openFDA_no_api_key",
    call = call
  )
```

could be, for conciseness and for making the URL clickable,

```r
  cli::cli_abort(
    c("To use {.pkg openFDA}, you must set an openFDA API key.",
      "i" = "Go to {.url https://open.fda.gov/apis/authentication/}
            to get an openFDA API key, then supply it to
            {.fun set_api_key} to cache it for use in this session."
    ),
    class = "openFDA_no_api_key",
    call = call
  )
```
- More generally there's no need to use `paste0()` within cli calls.

- While I understand the advantages of the function's returning an httr2 object, I wonder whether your package should also offer a bit more convenience for users that know less about httr2 details. This doesn't need to be settled now unless you want to: the opinion of reviewers will be useful.

- In the main vignette, in "By default, openFDA() will ask you if you want to perform paging, but this tends to break vignettes." you might want to clarify what you mean by performing paging: fetching all pages there are without the user's having to worry about paging.

- Regarding paging, rather than having a special argument, I like the approach of the `gh::gh()` function when you write `.limit = .Inf` to get all results.

- If the users deal with paging themselves, how do they enter which page they want to request?

- "We can suppress this warning by setting paging to "no-quiet"" is not intuitive for me because I read it as "not quiet", that is, "verbose". :smile_cat: More on verbosity control: <https://ropensci.org/blog/2024/02/06/verbosity-control-packages/>

- " The last date at which this openFDA endpoint was updated." made me wonder whether your package should support some sort of caching?

- I wonder whether the meaning of the `count` argument would be clearer if it were called `count_by` or `count_on`.

- Reading "So, for example, a search for drugs with the class "thiazide diuretic" should be formatted as "openfda.pharm_class_epc:\"thiazide diuretic\""" made me wonder whether the package could provide some helpers to build queries. Then I saw there are such helpers, could they already be mentioned in this example?

- Would it make any sense to make it possible to write queries a bit more like in R

```r
search_term <- format_search_term(c("openfda.generic_name" = "amoxicillin",
                                    "openfda.route" = "oral"),
                                  mode = "and")
```

would be

```r
search_term <- format_search_term(openfda.generic_name == "amoxicillin",
                                   & openfda.route == "oral")
```

Similarly `format_sort_term(c("openfda.generic_name" = "asc"))` could be something like `format_sort_term(asc(openfda.generic_name))`.

(for instance the [glitter package](https://lvaudor.github.io/glitter/) supports a domain-specific language for SPARQL, but then SPARQL is more complicated than what the FDA API requires)

- In any case it'd be nice not to have to use `c()` in `format_search_term()`, you could achieve that by using `...`and having other arguments' names start with a dot, like in `gh::gh()`.

- Wrong format "[httr2]" in https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/vignettes/openFDA.Rmd#L282.

- The intro of https://simpar1471.github.io/openFDA/articles/using_openFDA.html should explain what BNF is.

- The font you currently use for code has ligatures, which might be puzzling for some readers when they see `|>` as an arrow head.

- [`if (!inherits(warn_on_http_error, "logical")) {`](https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/check_args.R#L10-L11) should be `if (!is.logical(warn_on_http_error)) {`.

- [`if (!inherits(sort, "character")) {`](https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/check_args.R#L93-L94) should be `if (!is.character(sort)) {`. This comes up again at least in https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/check_args.R#L139-L140.

- On https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/check_args.R#L167-L168, the condition should use `is.numeric()`. I'd recommend going through all occurrences of `inherits()` to see whether there's a base R function covering the case.

- [`if (param < 1 & vname != "limit") {`](https://github.com/simpar1471/openFDA/blob/e2f59d7e0d65e28bdef601696f615580db043488/R/check_args.R#L184) should be `if (param < 1 && vname != "limit") {`. You could run this linter: https://lintr.r-lib.org/reference/vector_logic_linter.html. (From a cursory look most conditionals use the right pattern)

- I'm impressed by your using and testing error classes. :ok_hand:

- For testing why not use some sort of response caching through [{vcr}](https://docs.ropensci.org/vcr/) (check out its development version) or [{httptest2}](https://enpiar.com/httptest2/)?

- Are the tests calling the API not skipped on CRAN or did I miss something?

---
