### 2025-07-04 8771f71
docs: Change pkgdown code font away from one with ligatures
