### 2025-07-04 94cf79d
chore: drop `paste0()` from cli use; add {.url } instead of <url>
