### 2025-07-04 68d94b8
chore: fix lintr issue from `api_key == ""` using `nzchar()`
