### 2025-07-04 3bdfcbd
chore: replace unnecessary `inherits()` calls (e.g. `inherits(x, "character")`) with simpler `is.****()` calls
