### 2025-07-04 7c35388
docs: update DESCRIPTION to better describe openFDA reference
