+++ DESCRIPTION
@@ -8,8 +8,9 @@ Authors@R:
 Description: The 'openFDA' API facilitates access to U.S. Food and Drug 
     Administration (FDA) data on drugs, devices, foodstuffs, tobacco, and more.
     This package makes the API easily accessible with 'httr2', returning
-    objects which the user can convert to JSON data and parse. Kass-Hout TA, Xu
-    Z, Mohebbi M et al. (2016) <doi:10.1093/jamia/ocv153>.
+    objects which the user can convert to JSON data and parse. The API is
+    described in: Kass-Hout TA, Xu Z, Mohebbi M et al. (2016)
+    <doi:10.1093/jamia/ocv153>.
 License: GPL (>= 3)
 URL: https://github.com/simpar1471/openFDA,
     https://simpar1471.github.io/openFDA/
