### 2025-07-04 9fc2b64
fix: replace bad `foo & bar` with `foo && bar` where appropriate
