### 2025-07-04 f031d11
docs: Add example to openFDA-package.R
