### 2025-07-04 51b7976
docs: add better dash to README
