+++ README.Rmd
@@ -26,7 +26,7 @@ knitr::opts_chunk$set(
 
 openFDA makes querying the [openFDA API](https://open.fda.gov/apis/) from R a breeze.
 The API itself serves publicly available data from the FDA about foods, drugs, devices, and more.
-This data includes data such as recall enforcement reports, adverse events, manufacturer details, and - again - even more!
+This data includes data such as recall enforcement reports, adverse events, manufacturer details, and -- again -- even more!
 Note that the data on openFDA has not been validated for clinical or production use.
 
 ## Installation
@@ -51,11 +51,12 @@ The full documentation for the API is online, so look at the [openFDA
 website](https://open.fda.gov/apis/) to get a full feel for the API itself.
 
 Before using the package, you will need to generate one at the [openFDA webpage](https://open.fda.gov/apis/authentication/).
-With an API key you can make 240 requests per minute, and are limited to 120,000 requests per day.
+With an API key you can make 240 requests per minute, and are limited to 120,000 requests per day (the package will handle this request rate throttling for you).
 Without a key, you can only make 1000 requests a day.
 For this package, if you try to run `openFDA()` without setting an API key, it will fail:
 
 ```{r store_api_key, echo = FALSE, eval = TRUE}
+rlang::push_options("openFDA.paging" = "no-quiet")
 my_hidden_api_key <- get_api_key()
 set_api_key(api_key = "")
 ```
