### 2025-07-04 720ef1e
docs: improve documentation for construct_search_term() in main vignette
