### 2025-07-04 7ed99dc
docs: be a bit clearer about what the BNF is
