### 2025-07-04 183c4a8
docs: add a note in main vignette about 'performing paging'
