### 2025-07-04 503ccbb
docs: fix formatting of httr2 in openFDA main vignette
