### 2025-07-05 0c6c116
chore: drop unecessary `c()` calls in test-format_sort_term.R
