### 2025-07-05 10e95ed
refactor: Supply API key in `openFDA()` with req_auth_basic(). Prevents  API key appearing in URL when printing httr2 responses.
