### 2025-07-05 c1a3359
chore: Add class to message about paging information, and a unit test to look out for it
