### 2025-07-07 #714 @maelle
Thank you @simpar1471 for your work and response!

> what to put as return documentation - since the option isn't really a function that returns something. What would you advise?

You can actually write "Nothing" instead of writing nothing. 😆 Often I see something like "Nothing, called for its side-effects". https://github.com/search?q=%22%40return+Nothing%22+user%3Acran&type=code&ref=advsearch

> I think the keyring package is the best option - can I update the package/docs as required once reviewers are assigned?

If you already want to make the change, please do it before we assign reviewers so that they might review the better version (and have an easier time doing the set up on their machine). It's fine if we need to put the submission on hold for a short while.

> I'll also wait for package review before adding helpers for the returned objects, but I can see that being useful for end-users (at least some wrappers around httr2::resp_body_json()).

Good idea to wait for more feedback as this is a bit more complicated.

> Based on the blog post you attached, it looks like best practice to split the verbosity and paging control up.

If you agree that's also something you could do before I look for reviewers.

> I'm not sure if it should support a caching system like memoisation, as it is possible (though very unlikely) that updates to the API could occur between calls, and memoisation would prevent the new results from being retrieved. Do you have a link to a package with an example of this functionality? I've checked some of the rOpenSci data access packages but not found a setup to refer to.

- Memoization with memoise could be per-session.
- httr2 has https://httr2.r-lib.org/reference/req_cache.html
- Sometimes the API does not have the caching headers but you know (from the API docs or experience) the data is only sometimes updated, so you could be safe updating only every 3 months for instance. You could use an approach like in comtradr https://github.com/ropensci/comtradr/blob/f5b2715bf6dddcaff039c3613e8f4ecb0607e5c0/R/zzz.R#L58 or nflreadr https://github.com/nflverse/nflreadr/blob/main/R/zzz.R (This was a topic in the rOpenSci slack workspace, in the package-maintenance channel, just last week, if you want to take a look there -- assuming you still have access from your first submission, please tell me if not)

It is possible that caching doesn't make much sense for your package's use cases, so you could skip this at this point.

> I agree that the argument count would be clearer if renamed to count_by or count_on, but named it count to match onto the openFDA API's field names, which makes me reluctant to change it.

Makes sense!

> Added a (too?) brief note about the BNF in https://github.com/simpar1471/openFDA/commit/7ed99dc3306eea147bb9f79f854350a20f83a859

No it's fine! For me the BNF is https://en.wikipedia.org/wiki/Biblioth%C3%A8que_nationale_de_France 😸

> I'd happily implement a vcr-enabled test suite over the process of the review. This could aid security, e.g. by supplying the API key under the Authorization header instead of including it in the request URLs. Edit: Moved API key to Auth header in https://github.com/simpar1471/openFDA/commit/10e95ed0da8683fa67f7565c9a01b0c7b3bf0961

I'd find it relevant if you took the time to add some HTTP testing before I look for reviewers, as reviewers might comment on this anyway so you and they might save time. I'm happy to help if needed. The change might be less work than you imagine and would make tests faster for when you're making other changes. One common difficulty is that on CI you need to provide a fake API key, that isn't used by the mocked requests, but that tricks your own package's checking that there's an API key. See https://books.ropensci.org/http-testing/httptest2.html for instance. Now, the test coverage is good, so you can disagree and I won't hold it against the submission. 😁
