### 2025-07-07 5c12a35
fix: add `@return` tag to openFDA-package.R
