### 2025-07-07 49cc75a
fix: add `@return` tag to openFDA-package.R
