### 2025-07-07 a483412
chore: add some minimum package versions in DESCRIPTION
