### 2025-07-07 2abb2e1
feat: Change API key internals - now use keyring for more secure credential storage
