### 2025-07-07 4d89dd1
fix: Better tests for new API key implementation
