### 2025-07-07 a8b95b9
feat: More options + tests for these! Added options `openFDA.paging_verbosity` and `openFDA.handle_http_errors`. The exact details for implementation are in NEWS.md.
