+++ R/openFDA.R
@@ -20,20 +20,6 @@
 #' @param skip A single integer describing how many records should be skipped.
 #'   If more records are skipped than are found in your search, the openFDA
 #'   API will return a 404 error.
-#' @param paging A single string describing whether results should be paged,
-#'   and whether relevant alerts should be printed to the console. By default,
-#'   `openFDA()` uses the string stored in `options("openFDA.paging")`, which
-#'   has the value `"ask"`. Permissible values include:
-#'
-#'   * `"ask"` - `openFDA()` will warn you that pagination is required and ask
-#'     if you want this to be done. Depending on user input, either a single
-#'     {httr2} response object or a list of {httr2} response objects will be
-#'     returned. This will throw an error if your R session is not in
-#'     interactive mode.
-#'   * `"always"` - `openFDA()` will always perform pagination. A list of
-#'     {httr2} responses will be returned when pagination occurs.
-#'   * `"never"` - `openFDA()` will never perform pagination. Only the first
-#'     {httr2} response will be returned.
 #' @param api_key A single-length character vector with your openFDA API key. By
 #'   default this is the result of `get_api_key()`. You must set this with
 #'   `set_api_key()`, or `openFDA()` will not work.
@@ -95,10 +81,55 @@
 #'
 #'   This argument is case-sensitive. By default, the package will target the
 #'   Drugs@FDA endpoint (`"drugs-drugsfda"`).
-#' @param warn_on_http_error A scalar logical value. If `TRUE` (the default),
-#'   common openFDA HTTP errors will cause explanatory warnings to be printed
-#'   If `FALSE`, the underlying httr2 response object will be returned with
-#'   no extra warnings.
+#' @param paging A single string describing whether results should be paged. You
+#'   can set this on a per-function call level with this parameter, or edit
+#'   the [package-level option][openFDA_options] `openFDA.paging` to set paging
+#'   behaviour across multiple `openFDA()` calls.
+#'
+#'   By default, `openFDA()` uses the string stored in
+#'   `options("openFDA.paging")`, which has the default value `"ask"`.
+#'   Permissible values for `paging` include:
+#'
+#'   * `"ask"` - `openFDA()` will warn you that pagination is required and ask
+#'     if you want this to be done. Depending on user input, either a single
+#'     {httr2} response object or a list of {httr2} response objects will be
+#'     returned. This will throw an error if your R session is not in
+#'     interactive mode.
+#'   * `"always"` - `openFDA()` will always perform pagination. A list of
+#'     {httr2} responses will be returned when pagination occurs.
+#'   * `"never"` - `openFDA()` will never perform pagination. Only the first
+#'     {httr2} response will be returned.
+#' @param paging_verbosity A single string describing whether messages about
+#'   pagination should be printed to the console. You can set this on a
+#'   per-function call level with this parameter, or edit the [package-level
+#'   option][openFDA_options] `openFDA.paging_verbosity` to set this behaviour
+#'   across multiple `openFDA()` calls.
+#'
+#'   By default, `openFDA()` uses the string stored in
+#'   `options("openFDA.paging_verbosity")`, which has the default value
+#'   `"verbose"`. Permissible values for `paging_verbosity` include:
+#'
+#'   * `"verbose"` - If paging can be performed, print a message to the console,
+#'     stating how many requests are required with a minimum estimate for the
+#'     time this will take.
[…]
