+++ R/zzz.R
@@ -1,7 +1,12 @@
 .onLoad <- function(libname, pkgname) {
-  rlang::push_options(openFDA.paging = "ask")
+  # openFDA <<- memoise::memoise(openFDA)
+  rlang::push_options(openFDA.paging = "ask",
+                      openFDA.paging_verbosity = "verbose",
+                      openFDA.handle_http_errors = "warn")
 }
 
 .onUnload <- function(libname, pkgname) {
-  rlang::push_options(openFDA.paging = NULL)
+  rlang::push_options(openFDA.paging = NULL,
+                      openFDA.paging_verbosity = "verbose",
+                      openFDA.handle_http_errors = NULL)
 }
