### 2025-07-07 2930bf4
chore: Minor improvement to `count` docs in `openFDA()` + remove sanitise_api_key()
