### 2025-07-07 1711400
fix: Add code to retrieve API key from env var in vignette/article - should be stored in GH secrets for CI runs
