### 2025-07-07 b28d202
fix: Add explicit key to main package vignette ONLY or it won't build
