### 2025-07-07 3e547a8
fix: Add explicit key to main package vignette ONLY or it won't build
