### 2025-07-07 1727805
fix: Don't try to use a keyring if not running interactively
