### 2025-07-07 39e3afa
fix: Don't try to use a keyring if not running interactively
