### 2025-07-07 1b573ab
fix: Fixed for issues in CI R CMD CHECK runs:
 * Bad .Rd links to httr2
 * Bad `get_`/`set_api_key()` examples
