+++ R/openFDA.R
@@ -92,13 +92,14 @@
 #'
 #'   * `"ask"` - `openFDA()` will warn you that pagination is required and ask
 #'     if you want this to be done. Depending on user input, either a single
-#'     {httr2} response object or a list of {httr2} response objects will be
-#'     returned. This will throw an error if your R session is not in
-#'     interactive mode.
+#'     [httr2](https://httr2.r-lib.org/) response object or a list of
+#'     [httr2](https://httr2.r-lib.org/) response objects will be returned. This
+#'     will throw an error if your R session is not in interactive mode.
 #'   * `"always"` - `openFDA()` will always perform pagination. A list of
-#'     {httr2} responses will be returned when pagination occurs.
+#'     [httr2](https://httr2.r-lib.org/) responses will be returned when
+#'     pagination occurs.
 #'   * `"never"` - `openFDA()` will never perform pagination. Only the first
-#'     {httr2} response will be returned.
+#'     [httr2](https://httr2.r-lib.org/) response will be returned.
 #' @param paging_verbosity A single string describing whether messages about
 #'   pagination should be printed to the console. You can set this on a
 #'   per-function call level with this parameter, or edit the [package-level
@@ -125,11 +126,13 @@
 #'   `"warn"`. Permissible values for `handle_http_errors` include:
 #'
 #'   * `"warn"`: If the returned HTTP code is not 200, issue a warning to the
-#'     console and return the underlying {httr2} response object.
+#'     console and return the underlying [httr2](https://httr2.r-lib.org/)
+#'     response object.
 #'   * `"error"`: If the returned HTTP code is not 200, throw an (informative)
 #'     error.
 #'   * `"silent"`: If the returned HTTP code is not 200, return the underlying
-#'     {httr2} response object without printing a warning.
+#'     [httr2](https://httr2.r-lib.org/) response object without printing a
+#'     warning.
 #' @examples
 #' if (httr2::secret_has_key("OPENFDA_KEY")) {
 #'   set_api_key(httr2::secret_decrypt(
