### 2025-07-07 20bdbcb
fix: Fixed more issues in examples for R CMD CHECK runs
