+++ man/openFDA.Rd
@@ -168,7 +168,8 @@ if (httr2::secret_has_key("OPENFDA_KEY")) {
   resp <- openFDA(search = "openfda.manufacturer_name:gilead*",
                   limit = 2,
                   skip = 10,
-                  paging = "no-quiet")
+                  paging = "never",
+                  paging_verbosity = "quiet")
 
   # The function returns an `httr2` object
   print(resp)
