### 2025-07-07 f8e09c2
fix: add interactive flag to some tests of keyring behaviour
