### 2025-07-07 #714 @simpar1471
Hi @maelle, some commits and comments re your points:

 - Added return documentation for package-level options in https://github.com/simpar1471/openFDA/commit/49cc75a636165229a20acb3413a914b056a05607
 - Added keyring support + unit tests in https://github.com/simpar1471/openFDA/commit/2abb2e17f714d047b8f256bc62b6ce7d28cffc70 and https://github.com/simpar1471/openFDA/commit/4d89dd1a42189a5abf535886fc4df491989d60b3
 - Added extra options for paging, paging verbosity (i.e. whether that warning should get printed), and what to do with returned non-`200` HTTP codes in https://github.com/simpar1471/openFDA/commit/a8b95b9aa087e53d5d74c3874f6a20dc38f8ffbe. This was a breaking change insofar as the `warn_on_http_error` argument in `openFDA()` was changed to `handle_http_errors` and now takes character input to logical input.
 - For caching, `httr2::req_cache()` won't work as the API returns a `Cache-Control` header with `no-cache`. The endpoints are updated at different rates, from daily to quarterly. As such, I'll memoise `openFDA()` in `.onLoad()` to at least keep things cached per-session. I'll add this after implementing the HTTP testing.
- Thanks for the pointer to the ropensci book on HTTP testing, I'll look to implement this next.
