### 2025-07-07 e70a965
chore: Add internal error to get_openFDA_options(), add more unit tests for options
