### 2025-07-07 8f426e5
refactor: Add rlang::signal() to HTTP errors and test for *class of the condition* rather than for errors/warnings specifically
