### 2025-07-07 8b8de73
refactor: Internal changes to error class names, some error message edits.
