+++ README.Rmd
@@ -52,7 +52,7 @@ library(openFDA)
 The full documentation for the API is online, so look at the [openFDA 
 website](https://open.fda.gov/apis/) to get a full feel for the API itself.
 
-Before using the package, you will need to generate one at the [openFDA webpage](https://open.fda.gov/apis/authentication/).
+Before using the package, you will need to generate an API key on the [openFDA website](https://open.fda.gov/apis/authentication/).
 With an API key you can make 240 requests per minute, and are limited to 120,000 requests per day (the package will handle this request rate throttling for you).
 Without a key, you can only make 1000 requests a day.
 For this package, if you try to run `openFDA()` without setting an API key, it will fail:
@@ -60,8 +60,16 @@ For this package, if you try to run `openFDA()` without setting an API key, it w
 ```{r store_api_key, echo = FALSE, eval = TRUE}
 rlang::push_options(openFDA.paging = "never",
                     openFDA.paging_verbosity = "quiet")
-my_hidden_api_key <- get_api_key()
-keyring::key_delete("OPENFDA_KEY", "openFDA", "openFDA")
+encrypted_api_key <- paste0("TEaDtqdFMq9_Montij5p9IY6T57IyqkbF8IYFVOpk-",
+                            "ttxotFUNdJSxgccAnkq4nQhplaf-r3deQ")
+decrypted_api_key <- httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY")
+suppressWarnings(try(
+  keyring::key_delete("OPENFDA_KEY", "openFDA", "openFDA"),
+  silent = TRUE)
+)
+suppressWarnings(try(
+  keyring::key_delete("OPENFDA_KEY", "openFDA"), silent = TRUE)
+)
 ```
 
 ```{r no_api_key, error = TRUE}
@@ -74,17 +82,18 @@ search <- openFDA(
 That's not good.
 To fix this error, get an API key from the webpage linked above.
 Once you have your API key, use `set_api_key()` to store your API key.
-Once you run this function, **openFDA** will 'know' your API key until the session ends.
-Let's try using `openFDA()` again, after setting an API key:
+Once you run this function, **openFDA** will 'know' your API key, even for future sessions:
 
-```{r setting_api_key_fake, echo = TRUE, eval = FALSE}
+```{r setting_api_key_fake, echo = TRUE, eval = TRUE}
 # n.b. this API key is a fake, just for demonstration
 set_api_key(api_key = "iHXeqlbXtlYEwX9MYLwyCBuiZxfn98Sj3oX2FNtSx")
 ```
 
-```{r setting_api_key_real, echo = FALSE, eval = TRUE}
-set_api_key(api_key = my_hidden_api_key)
+```{r setting_api_key_real, echo = FALSE, eval = TRUE, message = FALSE}
+set_api_key(api_key = decrypted_api_key)
 ```
+
+Let's try using `openFDA()` again, now that you have an API key:
 
 ```{r example1_query}
 search <- openFDA(
