### 2025-07-07 7f3fdd7
chore: bump dev version to 9002 to reflect spate of recent changes, update codemeta.json
