### 2025-07-07 6f6648e
chore: update codecov badge in README and update codemeta.json
