### 2025-07-08 #714 @maelle
If you choose vcr rather than httptest2, note that the book on HTTP testing does not cover its development version yet, but vcr's Get Started vignette has been updated.
