### 2025-07-08 45f92ce
docs: make better use of cli's hyperlinking to documentation for keyring functions, and improve API key roxygen docs
