### 2025-07-08 0eeaa4e
feat: Added `vcr` to `Suggests` and in unit tests for openFDA, as well as use in `openFDA()` roxygen docs
