+++ R/openFDA.R
@@ -134,21 +134,17 @@
 #'     [httr2](https://httr2.r-lib.org/) response object without printing a
 #'     warning.
 #' @examples
-#' if (httr2::secret_has_key("OPENFDA_KEY")) {
-#'   set_api_key(httr2::secret_decrypt(
-#'     "TEaDtqdFMq9_Montij5p9IY6T57IyqkbF8IYFVOpk-ttxotFUNdJSxgccAnkq4nQhplaf-r3deQ",
-#'     "OPENFDA_KEY"
-#'   ))
-#'
-#'   resp <- openFDA(search = "openfda.manufacturer_name:gilead*",
-#'                   limit = 2,
-#'                   skip = 10,
-#'                   paging = "never",
-#'                   paging_verbosity = "quiet")
-#'
-#'   # The function returns an `httr2` object
-#'   print(resp)
+#' \dontshow{
+#'   vcr::insert_example_cassette("openFDA_examples", package = "openFDA")
 #' }
+#' resp <- openFDA(search = "openfda.manufacturer_name:gilead*",
+#'                 limit = 2,
+#'                 skip = 10,
+#'                 paging = "never",
+#'                 paging_verbosity = "quiet")
+#'
+#' # The function returns an `httr2` object
+#' print(resp)
 #'
 #' # Bad inputs will cause informative errors - here, a bad API key is supplied
 #' try(
@@ -156,6 +152,9 @@
 #'           api_key = "BAD_API_KEY",
 #'           limit = 1)
 #' )
+#' \dontshow{
+#'   vcr::eject_cassette()
+#' }
 #' @references
 #' Kass-Hout TA, Xu Z, Mohebbi M, Nelsen H, Baker A, LEvine J, Johansen E,
 #' Bright RA. **OpenFDA: an innovative platform providing access to a wealth of
@@ -379,7 +378,7 @@ openFDA_err_400_msg <- function(resp) {
 #' @rdname openFDA_http_errors
 #' @noRd
 openFDA_err_403_msg <- function(resp) {
-  api_key <- httr2::url_parse(resp$url)$query$api_key
+  api_key <- rlang::env_get(rlang::caller_env(7), nm = "api_key")
   c("The openFDA API returned a 403 error.",
     "!" = "This usually means that an invalid {.var api_key} was used.",
     "i" = cli::format_inline("Your API key was {.val {api_key}}."))
