+++ tests/testthat/test-openFDA.R
@@ -4,6 +4,10 @@ rlang::push_options(openFDA.paging = "never",
                     openFDA.paging_verbosity = "quiet")
 
 test_that("openFDA can call its API", {
+# This test we skip on CRAN, but keep in CI --> tests that the openFDA API is
+# actually available with a very simple query
+test_that("openFDA can call its API and reach correct endpoints", {
+  skip_if_offline()
   suppressMessages(
    set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
   )
@@ -11,15 +15,10 @@ test_that("openFDA can call its API", {
   # Simple query
   resp <- openFDA(
     search = c("openfda.brand_name" = "ozempic"),
-    limit = 5
+    limit = 2
   )
   expect_s3_class(resp, "httr2_response")
-})
-
-test_that("All openFDA endpoints are valid", {
-  suppressMessages(
-   set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
-  )
+  expect_equal(httr2::resp_body_json(resp)$meta$results$limit, expected = 2)
 
   endpoints <- c("animalandveterinary-event", "drug-event",
                  "drug-label", "drug-ndc", "drug-enforcement",
@@ -42,168 +41,156 @@ test_that("All openFDA endpoints are valid", {
   })
 })
 
-test_that("openFDA paging is possible", {
-  suppressMessages(
-   set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
-  )
+test_that(
+  desc = "openFDA paging is possible with correct printing behaviour", {
+  vcr::use_cassette("test-paging", {
+    # Query which requires paging
+    resps <- openFDA(search = "openfda.generic_name:\"semaglutide\"",
+                     limit = 2,
+                     paging = "always",
+                     paging_verbosity = "quiet")
+    purrr::map(resps, \(resp) expect_s3_class(resp, "httr2_response"))
 
-  # Query which requires paging
-  resps <- openFDA(
-    search = "openfda.brand_name:o*",
-    limit = 300,
-    paging = "always"
-  )
-
-  purrr::map(resps, \(resp) expect_s3_class(resp, "httr2_response"))
+    # Query which requires paging
+    expect_message(
+      openFDA(search = "openfda.generic_name:\"semaglutide\"",
+              limit = 2,
+              paging = "never",
+              paging_verbosity = "verbose"),
+      class = "openfda_message_paging"
+    )
+    # Query which requires paging
+    expect_no_message(
+      openFDA(search = "openfda.generic_name:\"semaglutide\"",
+              limit = 2,
+              paging = "never",
+              paging_verbosity = "quiet"),
+      class = "openfda_message_paging"
+    )
+  })
 })
 
-test_that("openFDA paging information is printed", {
-  suppressMessages(
-   set_api_key(httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY"))
-  )
-
-  # Query which requires paging
-  expect_message(
-    openFDA(search = "openfda.brand_name:o*",
-            limit = 300,
-            paging = "never",
-            paging_verbosity = "verbose"),
-    class = "openfda_message_paging"
-  )
-
-  # Query which requires paging
-  expect_no_message(
-    openFDA(search = "openfda.brand_name:o*",
-            limit = 300,
-            paging = "never",
-            paging_verbosity = "quiet"),
-    class = "openfda_message_paging"
-  )
-})
-
+# Does not need vcr as error is thrown *before* any HTTP requests sent
 test_that(
[…]
