+++ man/api_key.Rd
@@ -65,6 +65,9 @@ lock the \code{"openFDA"} keyring.
 }
 
 \examples{
+\dontshow{
+  try(prior_api_key <- get_api_key(), silent = TRUE)
+}
 # Set your openFDA API key with `set_api_key()`
 api_key <- "example_api_key"
 set_api_key(api_key)
@@ -74,4 +77,7 @@ get_api_key()
 
 # An error will be thrown you try to set your API key to an empty string.
 try(set_api_key(""))
+\dontshow{
+  try(set_api_key(api_key = prior_api_key), silent = TRUE)
+}
 }
