### 2025-07-08 26a46f7
fix: Revert openFDA.R example to not use `vcr::insert_example_cassette()` as this is not available on CI
