+++ man/api_key.Rd
@@ -66,7 +66,7 @@ lock the \code{"openFDA"} keyring.
 
 \examples{
 \dontshow{
-  try(prior_api_key <- get_api_key(), silent = TRUE)
+  suppressMessages(try(prior_api_key <- get_api_key(), silent = TRUE))
 }
 # Set your openFDA API key with `set_api_key()`
 api_key <- "example_api_key"
@@ -78,6 +78,6 @@ get_api_key()
 # An error will be thrown you try to set your API key to an empty string.
 try(set_api_key(""))
 \dontshow{
-  try(set_api_key(api_key = prior_api_key), silent = TRUE)
+  suppressMessages(try(set_api_key(api_key = prior_api_key), silent = TRUE))
 }
 }
