+++ inst/_vcr/openFDA_examples.yml
@@ -1,1060 +1,0 @@
-http_interactions:
-- request:
-    method: GET
-    uri: https://api.fda.gov/drug/drugsfda.json?search=openfda.manufacturer_name%3Agilead%2A&limit=2&skip=10
-  response:
-    status: 200
-    headers:
-      access-control-allow-credentials: 'true'
-      access-control-allow-origin: '*'
-      age: '0'
-      cache-control: no-cache, no-store, must-revalidate
-      content-encoding: gzip
-      content-security-policy: default-src 'none'
-      content-type: application/json; charset=utf-8
-      date: Tue, 08 Jul 2025 20:24:09 GMT
-      etag: W/"8203-xq/Gw6f8IOlMftc+bs1syNirKMw"
-      link: <https://api.fda.gov/drug/drugsfda.json?search=openfda.manufacturer_name%3Agilead%2A&limit=2&skip=12>;
-        rel="next"
-      strict-transport-security:
-      - max-age=31536000;
-      - max-age=31536000; preload
-      vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      via: https/1.1 api-umbrella (ApacheTrafficServer [cMsSf ])
-      x-api-umbrella-request-id: cnteqslrd582g7iceigg
-      x-cache: MISS
-      x-content-type-options: nosniff
-      x-frame-options: deny
-      x-ratelimit-limit: '240'
-      x-ratelimit-remaining: '239'
-      x-vcap-request-id: a9a603a7-f463-4731-587c-bea9a1e2050e
-      x-xss-protection: 1; mode=block
-    body:
-      string: |-
-        {
-          "meta": {
-            "disclaimer": "Do not rely on openFDA to make decisions regarding medical care. While we make every effort to ensure that data is accurate, you should assume all results are unvalidated. We may limit or otherwise restrict your access to the API in line with our Terms of Service.",
-            "terms": "https://open.fda.gov/terms/",
-            "license": "https://open.fda.gov/license/",
-            "last_updated": "2025-07-02",
-            "results": {
-              "skip": 10,
-              "limit": 2,
-              "total": 30
-            }
-          },
-          "results": [
-            {
-              "submissions": [
-                {
-                  "submission_type": "SUPPL",
-                  "submission_number": "14",
-                  "submission_status": "AP",
-                  "submission_status_date": "20211007",
-                  "review_priority": "PRIORITY",
-                  "submission_class_code": "EFFICACY",
-                  "submission_class_code_description": "Efficacy",
-                  "submission_property_type": [
-                    {
-                      "code": "Orphan"
-                    }
-                  ],
-                  "application_docs": [
-                    {
-                      "id": "68954",
-                      "url": "https://www.accessdata.fda.gov/drugsatfda_docs/appletter/2021/210251Orig1s014ltr.pdf",
-                      "date": "20211012",
-                      "type": "Letter"
-                    },
-                    {
-                      "id": "68979",
-                      "url": "http://www.accessdata.fda.gov/drugsatfda_docs/label/2021/210251s014lbl.pdf",
-                      "date": "20211013",
-                      "type": "Label"
-                    }
-                  ]
-                },
-                {
-                  "submission_type": "SUPPL",
-                  "submission_number": "20",
-                  "submission_status": "AP",
-                  "submission_status_date": "20241008",
-                  "review_priority": "STANDARD",
-                  "submission_class_code": "MANUF (CMC)",
-                  "submission_class_code_description": "Manufacturing (CMC)",
-                  "application_docs": [
-                    {
-                      "id": "79957",
-                      "url": "http://www.accessdata.fda.gov/drugsatfda_docs/label/2024/210251Orig1s020lbl.pdf",
-                      "date": "20241022",
[…]
