### 2025-07-10 0af1885
fix: Add chunk of code which unlocks "openFDA" keyring in test-api_key.R (prevents failure because new keyring is locked)
