### 2025-07-10 7fe463e
docs: change CRAN badge type in README
