### 2025-07-10 c138d7d
fix: Add `skip_if_not_installed("vcr")` to skip vcr-dependent tests in `check-no-suggests` CI
