+++ tests/testthat/helper-vcr.R
@@ -1,5 +1,7 @@
-library("vcr") # *Required* as vcr is set up on loading
-invisible(vcr::vcr_configure(
-  dir = vcr::vcr_test_path("fixtures")
-))
-vcr::check_cassette_names()
+if ("vcr" %in% installed.packages()) {
+  library("vcr") # *Required* as vcr is set up on loading
+  invisible(vcr::vcr_configure(
+    dir = vcr::vcr_test_path("fixtures")
+  ))
+  vcr::check_cassette_names()
+}
