### 2025-07-10 115701a
tests: Add a variety of mocked tests to test-api_key.R for increased coverage in unit tests
