+++ tests/testthat/test-api_key.R
@@ -1,4 +1,4 @@
-# API key editing when keyrings are supported ----------------------------------
+# Set/get API key with keyrings ------------------------------------------------
 
 test_that("You can set and get an API key (with keyring support)", {
   # Only run on windows, as there's the Windows credential manager
@@ -8,46 +8,82 @@ test_that("You can set and get an API key (with keyring support)", {
   # develop on Windows. On GitHub, the env-var backend is user which does not
   # have keyring support
   skip_if_not(keyring::has_keyring_support())
-  # Create keyring for this test
-  try(
-    keyring::keyring_create(keyring = "openFDA", password = "password"),
-    silent = TRUE
-  )
-  try(
-    keyring::keyring_unlock(keyring = "openFDA", password = "password"),
-    silent = TRUE
-  )
 
   encrypted_api_key <- paste0("TEaDtqdFMq9_Montij5p9IY6T57IyqkbF8IYFVOpk",
                               "-ttxotFUNdJSxgccAnkq4nQhplaf-r3deQ")
   decrypted_key <- httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY")
+
+
+  # 1. In interactive mode, can a user create a new a keyring and then save the
+  #    API key? Use mocking to replicate interactive inputs.
+  with_mocked_bindings(
+    interactive = function() TRUE,
+    keyring_unlock = function(keyring) {
+      keyring::keyring_unlock(keyring = "openFDA", password = "password")
+    },
+    keyring_create = function(keyring) {
+      keyring::keyring_create(keyring = keyring, password = "password")
+    },
+    key_set = function(service, user, keyring, prompt) {
+      rlang::signal(message = prompt, class = "openfda_api_key_set_with_mock")
+      keyring::key_set_with_value(service, user, decrypted_key, keyring)
+    },
+    code = {
+      ## a. No keyring exists --> should issue messages
+      try(keyring::key_delete(service = "OPENFDA_KEY",
+                              username = "openFDA",
+                              keyring = "openFDA"),
+          silent = TRUE)
+      try(keyring::keyring_delete(keyring = "openFDA"), silent = TRUE)
+      expect_message(set_api_key(decrypted_key),
+                     class = "openfda_keyring_doesnt_exist") |>
+        expect_message(class = "openfda_api_key_set")
+
+      ## b. keyring exists but is locked --> should issue messages
+      keyring::keyring_lock(keyring = "openFDA")
+      expect_message(set_api_key(decrypted_key),
+                     class = "openfda_keyring_locked") |>
+        expect_message(class = "openfda_api_key_set")
+
+      ## c. keyring exists and is unlocked, no specific key provided --> should
+      ##    still set a new API key, check that right code reached with
+      ##    rlang::signal in mocked `key_set()`
+      expect_condition(set_api_key(),
+                       class = "openfda_api_key_set_with_mock") |>
+        expect_message(class = "openfda_api_key_set")
+    }
+  )
+
+  # 2. In non-interactive mode, can a user set the API key?
+  try(keyring::keyring_create(keyring = "openFDA", password = "password"),
+      silent = TRUE)
+  try(keyring::keyring_unlock(keyring = "openFDA", password = "password"),
+      silent = TRUE)
   # `set_api_key() should return NULL but issue a message
   expect_message(
     expect_null(set_api_key(decrypted_key)),
     class = "openfda_api_key_set"
   )
 
-  # Now check whether the stored API key is the same as this one
-  expect_equal(get_api_key(), decrypted_key)
+  # 3. Now check whether the stored API key is the same as this one
+  with_mocked_bindings(
+    interactive = function() TRUE,
+    code = {
[…]
