+++ NAMESPACE
@@ -5,3 +5,7 @@ export(format_sort_term)
 export(get_api_key)
 export(openFDA)
 export(set_api_key)
+importFrom(keyring,key_get)
+importFrom(keyring,key_set)
+importFrom(keyring,keyring_create)
+importFrom(keyring,keyring_unlock)
