### 2025-07-10 32a8305
chore: Minor mocking additions in test-openFDA.R
