+++ tests/fixtures/test-handle_http_errors.yml
@@ -14,25 +14,25 @@ http_interactions:
     headers:
       access-control-allow-credentials: 'true'
       access-control-allow-origin: '*'
-      age: '0'
+      age: '1'
       cache-control: no-cache, no-store, must-revalidate
       content-length: '123'
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Tue, 08 Jul 2025 19:37:05 GMT
+      date: Thu, 10 Jul 2025 11:46:56 GMT
       etag: W/"7b-YXSQm9yIUSoIv/rIi3ya3OZ8kH8"
       strict-transport-security:
       - max-age=31536000;
       - max-age=31536000; preload
       vary: Accept-Encoding
       via: https/1.1 api-umbrella (ApacheTrafficServer [cMsSf ])
-      x-api-umbrella-request-id: cnte5b677vf16btu9330
+      x-api-umbrella-request-id: cnugkfmfrj35r6dmihq0
       x-cache: MISS
       x-content-type-options: nosniff
       x-frame-options: deny
       x-ratelimit-limit: '240'
-      x-ratelimit-remaining: '239'
-      x-vcap-request-id: 0d9ed482-cd59-43ab-7009-d8504a4a0e2b
+      x-ratelimit-remaining: '209'
+      x-vcap-request-id: 935e1691-ff13-42bc-5a01-eb6dec807406
       x-xss-protection: 1; mode=block
     body:
       encoding: ''
@@ -44,8 +44,8 @@ http_interactions:
             "message": "Sorting allowed by non-analyzed fields only: invalid_field"
           }
         }
-  recorded_at: 2025-07-08 19:37:09 GMT
-  recorded_with: vcr/1.7.0, webmockr/2.0.0
+  recorded_at: 2025-07-10 11:47:00 GMT
+  recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
     uri: https://api.fda.gov/drug/drugsfda.json?search=openfda.brand_name:%22verapamil%22%22&limit=1
@@ -59,16 +59,16 @@ http_interactions:
       status_code: 403
       message: Forbidden
     headers:
-      date: Tue, 08 Jul 2025 19:37:05 GMT
+      date: Thu, 10 Jul 2025 11:46:56 GMT
       content-type: application/json
       access-control-allow-origin: '*'
       content-encoding: gzip
       strict-transport-security: max-age=31536000; preload
       vary: Accept-Encoding
-      x-api-umbrella-request-id: cnte5b8jtg0odium2ol0
+      x-api-umbrella-request-id: cnugkfpb4e6bmsr0rp50
       x-cache: MISS
       x-content-type-options: nosniff
-      x-vcap-request-id: 912552b2-4dc1-4ba7-58e4-f7cb7c0dca18
+      x-vcap-request-id: ca3d4bee-83af-471c-776a-74652defadd5
       x-xss-protection: 1; mode=block
       x-frame-options: DENY
     body:
@@ -81,8 +81,8 @@ http_interactions:
             "message": "An invalid api_key was supplied. Get one at https://api.fda.gov:443"
           }
         }
-  recorded_at: 2025-07-08 19:37:09 GMT
-  recorded_with: vcr/1.7.0, webmockr/2.0.0
+  recorded_at: 2025-07-10 11:47:00 GMT
+  recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
     uri: https://api.fda.gov/drug/drugsfda.json?search=openfda.brand_name:%223129084%22&limit=1
@@ -103,7 +103,7 @@ http_interactions:
[…]
