+++ tests/fixtures/test-paging.yml
@@ -19,9 +19,9 @@ http_interactions:
       content-encoding: gzip
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Tue, 08 Jul 2025 19:29:24 GMT
-      etag: W/"6417-f9qmpt+5Ge9FQ9OHdprE6lZpeVM"
-      link: <https://api.fda.gov/drug/drugsfda.json?search=openfda.generic_name%3A%22semaglutide%22&limit=2&skip=0&search_after=0%3DZKDg4pcBmHiQdPkOGjx->;
+      date: Thu, 10 Jul 2025 11:46:54 GMT
+      etag: W/"6417-2PYjMEIuFOY/KR0j0DMrIiXuIsQ"
+      link: <https://api.fda.gov/drug/drugsfda.json?search=openfda.generic_name%3A%22semaglutide%22&limit=2&skip=0&search_after=0%3DC-F-8pcBIMiVPoupJ-xf>;
         rel="next"
       strict-transport-security:
       - max-age=31536000;
@@ -30,13 +30,13 @@ http_interactions:
       - Accept-Encoding
       - Accept-Encoding
       via: https/1.1 api-umbrella (ApacheTrafficServer [cMsSf ])
-      x-api-umbrella-request-id: cnte1qoeulhcqeh2ak5g
+      x-api-umbrella-request-id: cnugkf9qr7m5i57pbjn0
       x-cache: MISS
       x-content-type-options: nosniff
       x-frame-options: deny
       x-ratelimit-limit: '240'
-      x-ratelimit-remaining: '239'
-      x-vcap-request-id: 24c1a74b-be72-42ec-4ed9-c3b8e3acc1da
+      x-ratelimit-remaining: '213'
+      x-vcap-request-id: 77ad73a8-3066-4ab2-7175-2e25072d3c33
       x-xss-protection: 1; mode=block
     body:
       encoding: ''
@@ -47,7 +47,7 @@ http_interactions:
             "disclaimer": "Do not rely on openFDA to make decisions regarding medical care. While we make every effort to ensure that data is accurate, you should assume all results are unvalidated. We may limit or otherwise restrict your access to the API in line with our Terms of Service.",
             "terms": "https://open.fda.gov/terms/",
             "license": "https://open.fda.gov/license/",
-            "last_updated": "2025-07-02",
+            "last_updated": "2025-07-09",
             "results": {
               "skip": 0,
               "limit": 2,
@@ -843,8 +843,8 @@ http_interactions:
             }
           ]
         }
-  recorded_at: 2025-07-08 19:29:25 GMT
-  recorded_with: vcr/1.7.0, webmockr/2.0.0
+  recorded_at: 2025-07-10 11:46:56 GMT
+  recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
     uri: https://api.fda.gov/drug/drugsfda.json?search=openfda.generic_name:%22semaglutide%22&limit=2
@@ -865,9 +865,9 @@ http_interactions:
       content-encoding: gzip
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Tue, 08 Jul 2025 19:29:24 GMT
-      etag: W/"6417-f9qmpt+5Ge9FQ9OHdprE6lZpeVM"
-      link: <https://api.fda.gov/drug/drugsfda.json?search=openfda.generic_name%3A%22semaglutide%22&limit=2&skip=0&search_after=0%3DZKDg4pcBmHiQdPkOGjx->;
+      date: Thu, 10 Jul 2025 11:46:54 GMT
+      etag: W/"6417-2PYjMEIuFOY/KR0j0DMrIiXuIsQ"
[…]
