### 2025-07-10 6a2b1f3
feat: `openFDA()` is now memoised, which will speed up repeated requests (if identical)
