### 2025-07-10 fe15034
fix: Fix CMD CHECK NOTE from accidentally change `username` to `user` in `key_set()` calls
