### 2025-07-10 ea770ad
fix: Fix CMD CHECK NOTE by moving sole vctrs package use to its own, non-memoised function
