### 2025-07-10 0238296
tests: More mocking in =API key functions, to test most/all of the control flow
