+++ R/openFDA-package.R
@@ -2,9 +2,13 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom keyring has_keyring_support
 #' @importFrom keyring key_get
 #' @importFrom keyring key_set
+#' @importFrom keyring key_set_with_value
 #' @importFrom keyring keyring_create
+#' @importFrom keyring keyring_is_locked
+#' @importFrom keyring keyring_list
 #' @importFrom keyring keyring_unlock
 ## usethis namespace: end
 NULL
