+++ tests/testthat/test-api_key.R
@@ -1,79 +1,76 @@
+# Set-up -----------------------------------------------------------------------
+
+## Use env-var backed; supported on all systems (and default on GitHub)
+options(keyring_backend = keyring::backend_env)
+
 # Set/get API key with keyrings ------------------------------------------------
 
 test_that("You can set and get an API key (with keyring support)", {
-  # Only run on windows, as there's the Windows credential manager
-  skip_on_os(c("mac", "linux", "solaris"))
-  options(keyring_backend = keyring::backend_wincred)
-  # Skip if no keyring support - I test a backend *with* keyring support as I
-  # develop on Windows. On GitHub, the env-var backend is user which does not
-  # have keyring support
-  skip_if_not(keyring::has_keyring_support())
-
-  encrypted_api_key <- paste0("TEaDtqdFMq9_Montij5p9IY6T57IyqkbF8IYFVOpk",
-                              "-ttxotFUNdJSxgccAnkq4nQhplaf-r3deQ")
-  decrypted_key <- httr2::secret_decrypt(encrypted_api_key, "OPENFDA_KEY")
-
-
-  # 1. In interactive mode, can a user create a new a keyring and then save the
-  #    API key? Use mocking to replicate interactive inputs.
+  # 1. In interactive mode, can a user create a new keyring and save the API
+  #    key? Mocking interactive() + keyring functions.
   with_mocked_bindings(
     interactive = function() TRUE,
-    keyring_unlock = function(keyring) {
-      keyring::keyring_unlock(keyring = "openFDA", password = "password")
-    },
+    has_keyring_support = function() TRUE,
+    keyring_list = function() list(keyring = "not-openFDA"),
     keyring_create = function(keyring) {
-      keyring::keyring_create(keyring = keyring, password = "password")
+      rlang::signal(message = keyring, class = "openfda_test_keyring_created")
     },
     key_set = function(service, username, keyring, prompt) {
-      rlang::signal(message = prompt, class = "openfda_api_key_set_with_mock")
-      keyring::key_set_with_value(service, username, decrypted_key, keyring)
+      rlang::signal(message = prompt, class = "openfda_test_api_key_set")
     },
     code = {
-      ## a. No keyring exists --> should issue messages
-      try(keyring::key_delete(service = "OPENFDA_KEY",
-                              username = "openFDA",
-                              keyring = "openFDA"),
-          silent = TRUE)
-      try(keyring::keyring_delete(keyring = "openFDA"), silent = TRUE)
-      expect_message(set_api_key(decrypted_key),
-                     class = "openfda_keyring_doesnt_exist") |>
+      set_api_key() |>
+        expect_null() |>
+        expect_message(class = "openfda_keyring_doesnt_exist") |>
+        expect_condition(class = "openfda_test_keyring_created") |>
+        expect_condition(class = "openfda_test_api_key_set") |>
         expect_message(class = "openfda_api_key_set")
+  })
 
-      ## b. keyring exists but is locked --> should issue messages
-      keyring::keyring_lock(keyring = "openFDA")
-      expect_message(set_api_key(decrypted_key),
-                     class = "openfda_keyring_locked") |>
-        expect_message(class = "openfda_api_key_set")
-
-      ## c. keyring exists and is unlocked, no specific key provided --> should
-      ##    still set a new API key, check that right code reached with
-      ##    rlang::signal in mocked `key_set()`
-      expect_condition(set_api_key(),
-                       class = "openfda_api_key_set_with_mock") |>
-        expect_message(class = "openfda_api_key_set")
-    }
-  )
-
[…]
