+++ NAMESPACE
@@ -5,7 +5,11 @@ export(format_sort_term)
 export(get_api_key)
 export(openFDA)
 export(set_api_key)
+importFrom(keyring,has_keyring_support)
 importFrom(keyring,key_get)
 importFrom(keyring,key_set)
+importFrom(keyring,key_set_with_value)
 importFrom(keyring,keyring_create)
+importFrom(keyring,keyring_is_locked)
+importFrom(keyring,keyring_list)
 importFrom(keyring,keyring_unlock)
