### 2025-07-11 1250427
chore: Add rOpenSci badge to README
