### 2025-07-11 2361789
chore: text in message for locked keyring
