### 2025-07-11 d39c5e8
feat: Move options docs to `openFDA_options()`, a new function which can be used to get/set package-level options with added checks on inputs + returned values
