+++ tests/testthat/test-options.R
@@ -2,38 +2,45 @@ rlang::push_options(openFDA.paging = "ask",
                     openFDA.paging_verbosity = "verbose",
                     openFDA.handle_http_errors = "warn")
 
-test_that("You can get/set options", {
-  expect_equal(rlang::peek_option("openFDA.paging"),
-               "ask")
-  expect_equal(rlang::peek_option("openFDA.handle_http_errors"),
-               "warn")
+test_that("`current_openFDA_options()` functions as desired", {
+  ## Successful if all options are present
+  checkmate::expect_character(current_openFDA_options(),
+                              len = length(valid_openFDA_options),
+                              any.missing = FALSE)
 
-  rlang::push_options(openFDA.paging = "always")
-  expect_equal(rlang::peek_option("openFDA.paging"), "always")
-})
-
-test_that(desc = "Internal code for checking options throws the right errors", {
-  expect_error(
-    get_openFDA_option(option_name = "bad-option-name"),
-    class = "openfda_invalid_option_name"
+  ## Single bad option
+  rlang::with_options(
+    openFDA.paging = "TEST-BAD-VALUE",
+    expect_error(current_openFDA_options(),
+                 class = "openFDA_bad_option_found",
+                 pattern = "Set a valid value for this option"),
   )
 
-  for (option in names(valid_openFDA_options)) {
-    valid_option <- sample(valid_openFDA_options[[option]], size = 1)
+  ## Multiple bad options
+  rlang::with_options(
+    openFDA.paging = "no",
+    openFDA.paging_verbosity = "LOUD!",
+    expect_error(current_openFDA_options(),
+                 class = "openFDA_bad_option_found",
+                 pattern = "Set valid values for these options")
+  )
+})
 
-    list(valid_option) |>
-      setNames(option) |>
-      options()
-    expect_equal(get_openFDA_option(option), valid_option)
+test_that("`openFDA_options()` reports options if none supplied", {
+  expect_message(openFDA_options(), class = "openFDA_report_current_options")
+})
 
-    list(paste0(valid_option, "-not_valid_anymore")) |>
-      setNames(option) |>
-      options()
-    expect_error(
-      get_openFDA_option(option),
-      class = "openfda_invalid_option_value"
-    )
-  }
+test_that("`openFDA_options()` can set new options", {
+  expect_message(openFDA_options(paging = "always"),
+                 class = "openFDA_options_set")
+})
+
+test_that("`openFDA_options()` errors with bad option values", {
+  expect_error(
+    openFDA_options(paging = "always", paging_verbosity = "bad",
+                    handle_http_errors = "silent"),
+    class = "openFDA_options_setting_bad_value"
+  )
 })
 
 # Set options back to default values before other tests
