+++ R/options.R
@@ -1,38 +1,207 @@
-# Get openFDA package-level options (internal) ---------------------------------
+# Get/set openFDA package-level options ----------------------------------------
 
-#' Retrieve and check an openFDA package-level option (internal)
-#' @param option_name A single string denoting an openFDA package-level option
-#' @param call A function call to use if error message required.
-#' @noRd
-get_openFDA_option <- function(option_name, call = rlang::caller_env()) {
-  if (!option_name %in% names(valid_openFDA_options)) {
-    cli_valid_openFDA_option_names <- cli::cli_vec(
-      names(valid_openFDA_options), style = list("vec-last" = " or ")
+
+#' Get and set openFDA package-level options
+#' @description User-configurable options used by openFDA, which provide a
+#' mechanism for setting default behaviours.
+#' @param paging A single string used to set the `openFDA.paging` option. The
+#'   default, `NULL`, indicates that you don't want to change its value. The
+#'   `openFDA.paging` option controls how `openFDA()` behaves when there are
+#'   enough results that [paging is
+#'   required](https://open.fda.gov/apis/paging/), and has the default value
+#'   `"ask"`. New values must be one of:
+#'
+#'   * `"ask"` - `openFDA()` will warn you that pagination is required and ask
+#'     if you want this to be done. Depending on user input, either a single
+#'     [httr2](https://httr2.r-lib.org/) response object or a list of
+#'     [httr2](https://httr2.r-lib.org/) response objects will be returned. This
+#'     will throw an error if your R session is not in interactive mode. When s
+#'     etting this option, the value must be one of:
+#'   * `"always"` - `openFDA()` will always perform pagination. A list of
+#'     [httr2](https://httr2.r-lib.org/) responses will be returned when
+#'     pagination occurs.
+#'   * `"never"` - `openFDA()` will never perform pagination. Only the first
+#'     [httr2](https://httr2.r-lib.org/) response will be returned.
+#' @param paging_verbosity A single string used to set the
+#'   `openFDA.paging_verbosity` option. The default, `NULL`, indicates that you
+#'   don't want to change its value. The `openFDA.paging_verbosity` option
+#'   controls whether `openFDA()` prints messages when paging is required, and
+#'   has the default value `"warn"`. New values must be one of:
+#'
+#'   * `"verbose"` - If paging can be performed, print a message to the console,
+#'     stating how many requests are required with a minimum estimate for the
+#'     time this will take.
+#'   * `"quiet"` - Print no messages to the console regarding paging.
+#' @param handle_http_errors A single string used to set the
+#'   `openFDA.handle_http_errors` option. The default, `NULL`, indicates that
+#'   you don't want to change its value. The `openFDA.handle_http_errors` option
+#'   controls how `openFDA()` reacts when responses from the openFDA API have
+#'   HTTP codes other than `200 OK`, and has the default value `"warn"`. New
+#'   values must be one of:
+#'
+#'   * `"warn"`: If the returned HTTP code is not 200, issue a warning to the
+#'   console and return the underlying [httr2](https://httr2.r-lib.org/)
+#'   response object.
+#'   * `"error"`: If the returned HTTP code is not 200, throw an (informative)
+#'   error.
+#'   * `"silent"`: If the returned HTTP code is not 200, return the underlying
+#'   [httr2](https://httr2.r-lib.org/) response object without printing a
+#'   warning.
+#' @examples
+#' # Print current options by leaving all input args as `NULL`:
+#' openFDA_options()
+#'
[…]

[…]
