### 2025-07-11 b178fe7
tets: Bolster coverage of `openFDA()` with more mocked tests
