+++ tests/fixtures/test-handle_http_errors.yml
@@ -19,20 +19,20 @@ http_interactions:
       content-length: '123'
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Thu, 10 Jul 2025 16:14:00 GMT
+      date: Fri, 11 Jul 2025 11:39:23 GMT
       etag: W/"7b-YXSQm9yIUSoIv/rIi3ya3OZ8kH8"
       strict-transport-security:
       - max-age=31536000;
       - max-age=31536000; preload
       vary: Accept-Encoding
       via: https/1.1 api-umbrella (ApacheTrafficServer [cMsSf ])
-      x-api-umbrella-request-id: cnukenqn1i9mugea8uv0
+      x-api-umbrella-request-id: cnv546uvr4spuk8vl1t0
       x-cache: MISS
       x-content-type-options: nosniff
       x-frame-options: deny
       x-ratelimit-limit: '240'
-      x-ratelimit-remaining: '239'
-      x-vcap-request-id: 3e7e432f-fc2d-4a7c-63e1-3ee059fb6dfd
+      x-ratelimit-remaining: '179'
+      x-vcap-request-id: baf3828d-140e-475a-7a0c-5447e338a97f
       x-xss-protection: 1; mode=block
     body:
       encoding: ''
@@ -44,7 +44,7 @@ http_interactions:
             "message": "Sorting allowed by non-analyzed fields only: invalid_field"
           }
         }
-  recorded_at: 2025-07-10 16:14:10 GMT
+  recorded_at: 2025-07-11 11:39:28 GMT
   recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
@@ -59,16 +59,16 @@ http_interactions:
       status_code: 403
       message: Forbidden
     headers:
-      date: Thu, 10 Jul 2025 16:14:00 GMT
+      date: Fri, 11 Jul 2025 11:39:23 GMT
       content-type: application/json
       access-control-allow-origin: '*'
       content-encoding: gzip
       strict-transport-security: max-age=31536000; preload
       vary: Accept-Encoding
-      x-api-umbrella-request-id: cnukent7s5t9amffph90
+      x-api-umbrella-request-id: cnv54719ni7adklbn040
       x-cache: MISS
       x-content-type-options: nosniff
-      x-vcap-request-id: c00a9123-af98-47f8-6c81-98878632e621
+      x-vcap-request-id: fcda5a41-b4ce-4918-530e-b2165b999123
       x-xss-protection: 1; mode=block
       x-frame-options: DENY
     body:
@@ -81,60 +81,11 @@ http_interactions:
             "message": "An invalid api_key was supplied. Get one at https://api.fda.gov:443"
           }
         }
-  recorded_at: 2025-07-10 16:14:10 GMT
+  recorded_at: 2025-07-11 11:39:28 GMT
   recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
     uri: https://api.fda.gov/drug/drugsfda.json?search=openfda.brand_name:%223129084%22&limit=1
-    body:
-      encoding: ''
-      string: ''
-    headers:
-      Authorization: Basic MTVsaTgxWkU3bXBlWkJUQ2tEMjBaZ0xnYUdnV1RkQmV1bWlJMnF4cjo=
-  response:
-    status:
-      status_code: 404
-      message: Not Found
-    headers:
-      access-control-allow-credentials: 'true'
-      access-control-allow-origin: '*'
-      age: '0'
-      cache-control: no-cache, no-store, must-revalidate
-      content-encoding: gzip
-      content-security-policy: default-src 'none'
-      content-type: application/json; charset=utf-8
[…]
