+++ NAMESPACE
@@ -15,3 +15,4 @@ importFrom(keyring,keyring_create)
 importFrom(keyring,keyring_is_locked)
 importFrom(keyring,keyring_list)
 importFrom(keyring,keyring_unlock)
+importFrom(utils,menu)
