### 2025-07-11 9c51c46
docs: Improve `openFDA()` docs
