### 2025-07-11 f132b3c
fix: Change `pattern` to `regex` in `expect_condition()` in test-options.R
