### 2025-07-22 66c975e
chore: update http test .yml files
