+++ tests/fixtures/test-handle_http_errors.yml
@@ -19,20 +19,20 @@ http_interactions:
       content-length: '123'
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Fri, 11 Jul 2025 11:39:23 GMT
+      date: Tue, 22 Jul 2025 16:26:50 GMT
       etag: W/"7b-YXSQm9yIUSoIv/rIi3ya3OZ8kH8"
       strict-transport-security:
       - max-age=31536000;
       - max-age=31536000; preload
       vary: Accept-Encoding
       via: https/1.1 api-umbrella (ApacheTrafficServer [cMsSf ])
-      x-api-umbrella-request-id: cnv546uvr4spuk8vl1t0
+      x-api-umbrella-request-id: co6bqov7ua6fgua2765g
       x-cache: MISS
       x-content-type-options: nosniff
       x-frame-options: deny
       x-ratelimit-limit: '240'
-      x-ratelimit-remaining: '179'
-      x-vcap-request-id: baf3828d-140e-475a-7a0c-5447e338a97f
+      x-ratelimit-remaining: '190'
+      x-vcap-request-id: 955f2d66-df10-4591-48fc-5e14d0547ab1
       x-xss-protection: 1; mode=block
     body:
       encoding: ''
@@ -44,7 +44,7 @@ http_interactions:
             "message": "Sorting allowed by non-analyzed fields only: invalid_field"
           }
         }
-  recorded_at: 2025-07-11 11:39:28 GMT
+  recorded_at: 2025-07-22 16:26:54 GMT
   recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
@@ -59,16 +59,16 @@ http_interactions:
       status_code: 403
       message: Forbidden
     headers:
-      date: Fri, 11 Jul 2025 11:39:23 GMT
+      date: Tue, 22 Jul 2025 16:26:50 GMT
       content-type: application/json
       access-control-allow-origin: '*'
       content-encoding: gzip
       strict-transport-security: max-age=31536000; preload
       vary: Accept-Encoding
-      x-api-umbrella-request-id: cnv54719ni7adklbn040
+      x-api-umbrella-request-id: co6bqp1mh0sb3u3nm51g
       x-cache: MISS
       x-content-type-options: nosniff
-      x-vcap-request-id: fcda5a41-b4ce-4918-530e-b2165b999123
+      x-vcap-request-id: 378c1e80-f861-4f4c-55a4-d6a2a6143c85
       x-xss-protection: 1; mode=block
       x-frame-options: DENY
     body:
@@ -81,7 +81,7 @@ http_interactions:
             "message": "An invalid api_key was supplied. Get one at https://api.fda.gov:443"
           }
         }
-  recorded_at: 2025-07-11 11:39:28 GMT
+  recorded_at: 2025-07-22 16:26:54 GMT
   recorded_with: vcr/1.7.0, webmockr/2.1.0
 - request:
     method: get
@@ -98,12 +98,12 @@ http_interactions:
     headers:
       access-control-allow-credentials: 'true'
       access-control-allow-origin: '*'
-      age: '1'
+      age: '0'
       cache-control: no-cache, no-store, must-revalidate
       content-encoding: gzip
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Fri, 11 Jul 2025 11:39:24 GMT
+      date: Tue, 22 Jul 2025 16:26:51 GMT
       etag: W/"50-Gov3eVBQAYPDbwMSq8YtHKLjrHw"
       strict-transport-security:
       - max-age=31536000;
@@ -112,13 +112,13 @@ http_interactions:
       - Accept-Encoding
[…]
