### 2025-08-02 #714 @tanho63
**Briefly describe any working relationship you have (had) with the package authors.**

-   No prior relationship, was invited to review by @maelle

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

It is immediately evident that a good deal of thought and effort went into this package, nice work! A stream of feedback thoughts here:

-   I tend to agree with the editor comments that it would be convenient to parse the response automatically, especially since the API seems to always returns JSON. At one point in httr, Hadley [recommended](https://github.com/r-lib/httr/blob/2554ed25d290da83a31dbbe36afa3cb358ec34f9/vignettes/api-packages.Rmd#L155-L187) returning an S3 object as the HTTP response object, which allows you to attach the response as an attribute along with the parsed object. (I don't think Hadley has changed his opinion on this, just that it didn't make sense for the scope of httr2's vignettes). This also would allow for some nice print methods, e.g.

``` r
R> nflreadr::load_player_stats()
── nflverse player stats: offense ───────────────────────────────────────────────────────────
ℹ Data updated: 2025-05-07 04:12:49 EDT
# A tibble: 5,597 × 53
   player_id player_name player_display_name position position_group headshot_url recent_team
   <chr>     <chr>       <chr>               <chr>    <chr>          <chr>        <chr>
 1 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 2 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 3 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 4 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 5 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 6 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 7 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 8 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
 9 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
10 00-00234… A.Rodgers   Aaron Rodgers       QB       QB             https://sta… NYJ
# ℹ 5,587 more rows
# ℹ 46 more variables: season <int>, week <int>, season_type <chr>, opponent_team <chr>,
#   completions <int>, attempts <int>, passing_yards <dbl>, passing_tds <int>,
#   interceptions <dbl>, sacks <dbl>, sack_yards <dbl>, sack_fumbles <int>,
#   sack_fumbles_lost <int>, passing_air_yards <dbl>, passing_yards_after_catch <dbl>,
#   passing_first_downs <dbl>, passing_epa <dbl>, passing_2pt_conversions <int>, pacr <dbl>,
#   dakota <dbl>, carries <int>, rushing_yards <dbl>, rushing_tds <int>, …
# ℹ Use `print(n = ...)` to see more rows
```

(or something more optimized for a list e.g. str(x, max.level = 2))

-   If you do go this route, you might also consider using RcppSimdJson or yyjsonr to parse the text body more performantly than the default jsonlite implementation in httr2.

-   I like the addition of memoising! You may also consider a bit more of a robust implementation by adding a default timeout, the option to cache to disk instead of memory, and the ability to turn caching off. Here's an example [implementation](https://github.com/nflverse/nflreadr/blob/main/R/zzz.R), although I would probably use `tools::R_user_dir()` now instead of the rappdirs package.

-   I think you're already on CRAN so you won't run into CRAN complaining about the superassignment in your memoise call, but if you run into problems with that in the future you can use assign + `rlang::ns_env()`: https://github.com/r-lib/memoise/issues/76#issuecomment-1662885970

-   I don't think you need to have [this chunk](https://github.com/simpar1471/openFDA/blob/main/R/zzz.R#L11-L15) to remove options onUnload, those options are scoped to this package so we may as well keep them there.

-   This section of [experimental/benchmark code](https://github.com/simpar1471/openFDA/blob/main/R/zzz.R#L21-L61) should probably live somewhere else aside from being bundled in the package, i.e. in a `dev/` folder that is .Rbuildignored or perhaps just in `data-raw`.

-   The [user agent and throttle rate](https://github.com/simpar1471/openFDA/blob/main/R/openFDA.R#L189-L193) should probably be customizable via R options rather than hardcoded.

-   I appreciate all the effort in error handling! Am curious about the motivation vs using default httr2 error printing.

-   Getting a warning on `devtools::load_all()` that `vcr::check_cassette_names()` has been deprecated in vcr 2.0.0, and on reading the changelog it seems that the `library(vcr)` call is no longer be needed. These seem to be brand new as of July 28th so totally understand why those changes are still here :P You may want to set a minimum vcr of 2.0.0 after updating to fix it. (Changelog here: https://recology.info/2025/07/vcr-v2/#lifecycle-changes)

-   re: \> Without a key, you can only make 1000 requests a day. For this package, if you try to run openFDA() without setting an API key, it will fail

    Does it make sense to actually allow running openFDA without a key if it will work for 1000 requests per day? How often would someone max out the request limit personally?

-   As a general code maintenance comment, I'm a fan of Jenny Bryan's thoughts [here](https://github.com/jennybc/code-smells-and-feels) about "never using `else`" in code, and instead trying to simplify if/else blocks as much as possible for readability. (This was mostly a response to reading through the api key section while debugging the below issue)

-   I believe there's an issue with the keyring setup:

``` r
set_api_key
keyring::key_list(keyring = "openFDA")
#>       service username
#> 1 OPENFDA_KEY  openFDA
get_api_key()
#> Error:
#> ! No openFDA API key was found by `keyring::key_get()`.
#> ℹ To use openFDA, you must set an openFDA API key. Go to
#>   <https://open.fda.gov/apis/authentication/> to get an openFDA API key, then
#>   use set_api_key() to cache it for use in this and future sessions.
#> Show Traceback
nchar(keyring::key_get("OPENFDA_KEY", keyring = "openFDA", username = "openFDA"))
#> [1] 11
```

On further inspection, I believe it's because [`openFDA_keyring_exists()`](https://github.com/simpar1471/openFDA/blob/main/R/api_key.R#L198-L200) has a bug:

``` r
openFDA_keyring_exists <- function(openFDA_keyring = "openFDA") {
  has_keyring_support() && "openFDA" %in% keyring_list()
}
has_keyring_support()
#> [1] TRUE

"openFDA" %in% keyring_list()
#> [1] FALSE

keyring::keyring_list()
#>   keyring num_secrets locked
#> 1   Login          13  FALSE
#> 2 openFDA           0  FALSE

"openFDA" %in% keyring_list()$keyring
#> [1] TRUE
```

Fixing the bug locally and then proceeding with rest of the review.

-   `set_api_key()` seems to return a [message](https://github.com/simpar1471/openFDA/blob/66c975e59eea8c193452e5f89616656376ae29a6/R/api_key.R#L112-L117) suggesting to get the key via the `keyring::key_get()` function, should this suggest `openFDA::get_api_key()` instead?

-   `set_api_key()` has this [blank if case](https://github.com/simpar1471/openFDA/blob/66c975e59eea8c193452e5f89616656376ae29a6/R/api_key.R#L94-L96), what's this for?

-   As a general observation I like the effort to use the keyring package to enforce good habits on users, but I would probably not have brought in so much of the keyring handling into the package and narrowed the scope to just the API-specific parts by having the API key be an environment variable or an R object stored in the package environment itself (since it adds complexity and maintenance burden - you basically need to support all the parts of keyring going forward when you'd probably prefer to only support the openFDA API changes). To encourage usage of keyring, I might demonstrate it in the README or vignettes instead.

-   I like that you're thinking through the paging part for the user! I think calling the argument `limit` implies that I only want `limit` results, so being offered to return more results is unexpected. I think the behaviour I'd probably expect is that `limit` controls the final number of results returned (and can accept `Inf` to request all records), and then the page size is controlled as an option that defaults to / maxes out at 1000. This is a similar interface to `gh::gh()`.

-   Looking at the [behaviour of the paging option](https://github.com/simpar1471/openFDA/blob/main/R/options.R#L15-L17), strongly avoid returning different output types based on an R option being set or not - causes surprises if you receive a list when you expect an httr2 object and vice-versa. I think this goes back to the parsing discussion - once parsed, returning a list of api response content is much easier to combine in a sensible way (maybe automatically collating list-objects by name?)

-   re:

    > I've not made a DSL before, and could give it a try, but would like to hear back from reviewers as to whether they think this would be more useful than your base recommendation of dropping the need for c().

    I think either of these options are fine, the latter being an easier lift and probably sufficient. If you wanted to go this road I would also recommend changing your other argument names in `format_search_term()` to be prefixed with a dot, to mark them receiving special handling.

-   Tests seem appropriately comprehensive and run in an acceptable time, R CMD check passes without issues.

-   Vignette seems to work well.

-   Vignette has a lot of `#> Warning in default_backend_auto(): Selecting 'env' backend. Secrets are stored in environment variables` printouts. Consulting the keyring docs, I think you want to set

``` r
options(keyring_backend = "backend_env")
```

  in the [setup_hidden](https://github.com/simpar1471/openFDA/blob/main/vignettes/articles/using_openFDA.Rmd#L14-L19) chunk to silence this.

<summary>Reviewer Session Info</summary>

```
─ Session info ───────────────────────────────────────────────
 os       Ubuntu 22.04.5 LTS
 system   x86_64, linux-gnu
 ui       X11
 collate  en_CA.UTF-8
 ctype    en_CA.UTF-8
 tz       America/Toronto
 pandoc   2.9.2.1 @ /usr/bin/pandoc
 quarto   1.5.54 @ /opt/quarto/bin/quarto

─ Packages ───────────────────────────────────────────────────
 ! package     * version    date (UTC) lib source
   askpass       1.2.1      2024-10-04 [1] RSPM
   backports     1.5.0      2024-05-23 [1] RSPM
   brio          1.1.5      2024-04-24 [1] RSPM
   cachem        1.1.0      2024-05-16 [1] RSPM
   checkmate     2.3.2      2024-07-29 [1] RSPM
   cli           3.6.5      2025-04-23 [1] RSPM
   crayon        1.5.3      2024-06-20 [1] RSPM
   curl          6.4.0      2025-06-22 [1] RSPM
   desc          1.4.3      2023-12-10 [1] RSPM
   devtools    * 2.4.5      2022-10-11 [1] RSPM
   digest        0.6.37     2024-08-19 [1] RSPM
   ellipsis      0.3.2      2021-04-29 [1] RSPM
   fastmap       1.2.0      2024-05-15 [1] RSPM
   fs            1.6.6      2025-04-12 [1] RSPM
   glue          1.8.0      2024-09-30 [1] RSPM
   htmltools     0.5.8.1    2024-04-04 [1] RSPM
   htmlwidgets   1.6.4      2023-12-06 [1] RSPM
   httpuv        1.6.16     2025-04-16 [1] RSPM
   httr2         1.2.1      2025-07-22 [1] RSPM
   jsonlite      2.0.0      2025-03-27 [1] RSPM
   keyring       1.4.1      2025-06-15 [1] RSPM
   later         1.4.2      2025-04-08 [1] RSPM
   lifecycle     1.0.4      2023-11-07 [1] RSPM
   magrittr      2.0.3      2022-03-30 [1] RSPM
   memoise       2.0.1      2021-11-26 [1] RSPM
   mime          0.13       2025-03-17 [1] RSPM
   miniUI        0.1.2      2025-04-17 [1] RSPM
 P openFDA     * 0.1.0.9002 2025-07-16 [?] local
   openssl       2.3.3      2025-05-26 [1] RSPM
   pkgbuild      1.4.8      2025-05-26 [1] RSPM
   pkgload       1.4.0      2024-06-28 [1] RSPM
   prettyunits   1.2.0      2023-09-24 [1] RSPM
   profvis       0.4.0      2024-09-20 [1] RSPM
   promises      1.3.2      2024-11-28 [1] RSPM
   purrr         1.1.0      2025-07-10 [1] RSPM
   R6            2.6.1      2025-02-15 [1] RSPM
   rappdirs      0.3.3      2021-01-31 [1] RSPM
   Rcpp          1.0.14     2025-01-12 [1] RSPM
   remotes       2.5.0      2024-03-17 [1] RSPM
   rlang         1.1.6      2025-04-11 [1] RSPM
   rprojroot     2.1.0      2025-07-12 [1] RSPM
   rstudioapi    0.17.1     2024-10-22 [1] RSPM
   sessioninfo   1.2.3      2025-02-05 [1] RSPM
   shiny         1.10.0     2024-12-14 [1] RSPM
   testthat    * 3.2.3      2025-01-13 [1] RSPM
   urlchecker    1.0.1      2021-11-30 [1] RSPM
   usethis     * 3.1.0      2024-11-26 [1] RSPM
   vcr         * 2.0.0      2025-07-23 [1] RSPM
   vctrs         0.6.5      2023-12-01 [1] RSPM
   withr         3.0.2      2024-10-28 [1] RSPM
   xtable        1.8-4      2019-04-21 [1] RSPM

 [1] /home/tan/R/x86_64-pc-linux-gnu-library/4.5
 [2] /opt/R/4.5.0/lib/R/library

 * ── Packages attached to the search path.
 P ── Loaded and on-disk path mismatch.
```
