### 2025-08-05 #714 @nrennie
-   No prior relationship with author - invited to review by @maelle

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

-   [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

Overall, I think this package is a good contribution and is well-documented with some nice examples and tests. There’s also a lot of nice user-functionality e.g. to format requests properly which has obviously been thought about a lot. Most of my comments are relatively minor:

-   The vignette has an example with an explicit API key in the .Rmd file. Placing a (working) API key in plain text in a public GitHub repo isn’t a great idea - and I’d really suggest removing it - even though you have a statement of *This API key is only used for this vignette*. It also might get flagged by CRAN. Since vignettes are built locally, and only checked by CRAN, I would suggest adding a `NOT_CRAN` variable and evaluating conditionally. You can use GitHub secrets to render via GitHub Actions if need be.

-   It’s not clear why a user needs to set an API key since `Without a key, you can only make 1000 requests a day. For this package, if you try to run openFDA() without setting an API key, it will fail`. If the API itself allows some requests without a key, why does the package not allow it? Enabling this would also solve some of the issues of getting examples to run when no API key is present i.e. would fix the problem above.

-   I agree with comments from others that some functions to parse the response would be useful, and would really widen the user base. There’s likely a lot of users who are very familiar with the domain, but less confident using R (especially for parsing JSON which many won’t have done). Being able to get results directly in a data.frame would make it easier for them to see how they could benefit from using the package.

-   In the vignette, there’s a warning:

    #> Warning in default_backend_auto(): Selecting 'env' backend. Secrets are stored
    #> in environment variables

I’d suggest addressing the issue that causes the warning, and/or discussing why this warning occurs which would perhaps be useful to users who may get the same warning. Perhaps also add some examples of using `keyring` in a separate vignette specifically about setting up API keys would be useful. With many packages, I’ve found step 1 of API set up is the hardest part so having explicit instructions and common errors is really helpful.

#### Minor comments

-   I would put `CONTRIBUTING.md` in the root directory (rather than inside `.github/`) so that it’s easier to find and sits beside the code of conduct, but that’s perhaps more of a personal preference.

-   Perhaps state in the README why this package should be preferred over others that exist. The `openfda` package doesn’t appear to be actively maintained and hasn’t been updated in 10 years - maybe make it clearer that this is a different package completely, that just has a very similar name.

-   Consider adding a note about whether or not search terms are case sensitive.

-   When using dark mode on the package website, the plots in the examples aren’t really visible (black lines on dark grey background). Can you explicitly add a white background to plots?

-   <s>The paging option `"no-quiet"` is a bit confusing - I understand it means *no paging* and do it *quietly* but it reads as *don’t do it quietly*. I think it would be better to include two arguments - one for whether not paging is performed, and another argument for verbosity (e.g. `quiet = TRUE` would be more in keeping with existing packages).</s> Edit: I think this has been updated in the package, it’s just the [“Getting Started”](https://simpar1471.github.io/openFDA/articles/openFDA.html#large-result-sets) page that’s outdated.

<summary>
Reviewer Session Info
</summary>

    R version 4.5.0 (2025-04-11 ucrt)
    Platform: x86_64-w64-mingw32/x64
    Running under: Windows 11 x64 (build 26100)

      LAPACK version 3.12.1

    [1] LC_COLLATE=English_United Kingdom.utf8  LC_CTYPE=English_United Kingdom.utf8
    [3] LC_MONETARY=English_United Kingdom.utf8 LC_NUMERIC=C
    [5] LC_TIME=English_United Kingdom.utf8

    time zone: Europe/London
    tzcode source: internal

    [1] openFDA_0.1.0.9002

     [1] digest_0.6.37     R6_2.6.1          fastmap_1.2.0     xfun_0.52
     [5] cachem_1.1.0      knitr_1.50        memoise_2.0.1     htmltools_0.5.8.1
     [9] rmarkdown_2.29    lifecycle_1.0.4   cli_3.6.4         pak_0.9.0
    [13] rsconnect_1.5.0   pkgload_1.4.0     compiler_4.5.0    rstudioapi_0.17.1
    [17] tools_4.5.0       evaluate_1.0.4    yaml_2.3.10       rlang_1.1.6
    [21] keyring_1.4.1
