### 2026-05-03 #714 @maurolepore
I see @maelle is steering this through and the submission is on hold per your conversation in February. Please reach out whenever you're ready to resume.

Thanks for sharing your work with the rOpenSci community.
