### 2026-09-04 #714 @maelle
Just sent a new email.

Our policies state we only close on-hold issues after one year, and I really hope the submission can resume before then. 🤞
